+++ coffeeapp/src/app/_model/alive.model.ts
