+++ coffeeapp/src/app/_model/consumption.model.ts
