+++ coffeeapp/src/app/_model/index.ts
@@ -1,0 +1,5 @@
+export * from './alive.model';
+export * from './config.model';
+export * from './consumption.model';
+export * from './login.model';
+export * from './sensordata.model';
