+++ coffeeapp/src/app/_model/login.model.ts
@@ -1,0 +1,11 @@
+/**
+ * This class represents the result of the login.
+ */
+export class LoginModel {
+
+  success: boolean;
+
+  constructor() {
+  }
+
+}
