+++ coffeeapp/src/app/_model/sensordata.model.ts
